Implemented  project4.
Implemented MeshDrawer class, mvp matrix projection finder and VS and FS  for meshes.
When you add an obj file before adding texture at the beginning, you'll see the object black because ShowTexture is checked and no texture is loaded. After texture is added, you are able to see the object with texture. This is not a mistake, it is an implementation choice.